Modularize routes, use ExpressJS code structure
